+++ Kernel/Output/HTML/ToolBar/TicketProcess.pm
@@ -20,12 +20,17 @@ our @ObjectDependencies = (
     'Kernel::Output::HTML::Layout',
     'Kernel::System::Log',
     'Kernel::System::ProcessManagement::Process',
+    'Kernel::System::Ticket',
 );
 
 sub Run {
     my ( $Self, %Param ) = @_;
 
-    my $LogObject = $Kernel::OM->Get('Kernel::System::Log');
+    my $ConfigObject  = $Kernel::OM->Get('Kernel::Config');
+    my $LayoutObject  = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
+    my $LogObject     = $Kernel::OM->Get('Kernel::System::Log');
+    my $TicketObject  = $Kernel::OM->Get('Kernel::System::Ticket');
+    my $ProcessObject = $Kernel::OM->Get('Kernel::System::ProcessManagement::Process');
 
     for my $Needed (qw(Config)) {
         if ( !$Param{$Needed} ) {
@@ -40,10 +45,8 @@ sub Run {
     # check if frontend module is used
     my $Action = $Param{Config}->{Action};
     if ($Action) {
-        return if !$Kernel::OM->Get('Kernel::Config')->Get('Frontend::Module')->{$Action};
+        return if !$ConfigObject->Get('Frontend::Module')->{$Action};
     }
-
-    my $ProcessObject = $Kernel::OM->Get('Kernel::System::ProcessManagement::Process');
 
     my $ProcessList = $ProcessObject->ProcessList(
         ProcessState => ['Active'],
@@ -51,9 +54,32 @@ sub Run {
         Silent       => 1,
     );
 
-    return if ( !IsHashRefWithData($ProcessList) );
+    # prepare process list for ACLs, use only entities instead of names, convert from
+    #   P1 => Name to P1 => P1. As ACLs should work only against entities
+    my %ProcessListACL = map { $_ => $_ } sort keys %{$ProcessList};
 
-    my $LayoutObject = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
+    # validate the ProcessList with stored ACLs
+    my $ACL = $TicketObject->TicketAcl(
+        ReturnType    => 'Process',
+        ReturnSubType => '-',
+        Data          => \%ProcessListACL,
+        Action        => $Self->{Action},
+        UserID        => $Self->{UserID},
+    );
+
+    if ( IsHashRefWithData($ProcessList) && $ACL ) {
+
+        # get ACL results
+        my %ACLData = $TicketObject->TicketAclData();
+
+        # recover process names
+        my %ReducedProcessList = map { $_ => $ProcessList->{$_} } sort keys %ACLData;
+
+        # replace original process list with the reduced one
+        $ProcessList = \%ReducedProcessList;
+    }
+
+    return if ( !IsHashRefWithData($ProcessList) );
 
     # get item definition
     my $Text      = $LayoutObject->{LanguageObject}->Translate( $Param{Config}->{Name} );